requirements.txt
python-dotenv
openai
google-generativeai
pandas